requirements/requirements.txt
Django==1.11.15
django-cors-headers==2.5.2
django-extra-fields==1.0.0
django-filter==1.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
meld3==1.0.2
pbr==3.1.1
Pillow==5.4.1
psycopg2==2.7.5
psycopg2-binary==2.7.7
pycrypto==2.6.1
PyJWT==1.7.1
pytz==2018.5
six==1.10.0
slugify==0.0.1
stevedore==1.25.0
django-environ==0.4.5


python-slugify==1.2.6

gunicorn==19.9.0
django-extensions==2.1.4
django-autoslug

django-rest-framework-multi-slug-field>=0.2.0

argon2-cffi==20.1.0

drf-dynamic-fields
whitenoise==4.1.2

pydot==1.4.1 
pyparsing==2.4.6

drf-flex-fields==0.8.5

drf-yasg
drf-yasg[validation]